refactor: move test case to separate file
Related-to: #944
